+++ tests/Db/Expression/Value/Builder/ParamBuilderTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Tests\Db\Expression\Builder;
+namespace Yiisoft\Db\Tests\Db\Expression\Value\Builder;
 
 use PHPUnit\Framework\TestCase;
 use Yiisoft\Db\Expression\Value\Builder\ParamBuilder;